feat: :boom: add new exception subclasses and change exception constructors
This commit adds three new exception subclasses and changes the constructors of existing ones to match the Exception class. The new validations exceptions are ValidationError and FactoryError, and the new repository exceptions are AlreadyExistError, FetchError, NotFoundError, and PersistenceError. With these updates, our exceptions are more aligned with the base Exception class and can be more easily consumed by developers.