feat: :sparkles: add class event store to allow storing and retrieving domain events
This commit renames the `DomainEvent` class to `Event` and makes it an interface of `EventStore`, which now can store and retrieve domain events. The class `AggregateRoot` was modified to allow committing domain events to the event store.

The `Saga` interface was removed, and the `OptionalDateValidator` class had unused methods removed.