feat: ✨ add event factory interface and update aggregate root class
This commit adds the EventFactory interface to the codebase, allowing the creation of events. The AggregateRoot class is updated to implement the EventFactory interface in addition to the EventStore interface. The `createEvent` method is added to the AggregateRoot class, which creates and adds an event to the list of events to commit.